+++ lib/kagayaku_modules.dart
@@ -3,5 +3,6 @@ library kagayaku_modules;
 export 'src/kagayaku_modules_base.dart';
 export 'src/models/novel.dart';
 export 'src/models/module_info.dart';
+export 'src/models/novel_details.dart';
 export 'src/utils/get_source.dart';
 export 'src/utils/languages.dart';